requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
pyyaml
matplotlib
seaborn
jupyterlab